+++ src/components/stages/settings/Settings.jsx
@@ -1,21 +1,22 @@
 import React from 'react';
 import PropTypes from 'prop-types';
 
-const Settings = ({ renderModule }) => {
+const Settings = ({ displayLanguages, renderModule, translations }) => {
     if (!renderModule) {
         return null;
     }
 
     return (
         <div className={'settingsContainer'}>
-            <h1>Einstellungen</h1>
+            <h1>{translations.titleLabel}</h1>
 
-            <label>Anzeigesprache</label>
+            <label>{translations.displayLanguageLabel}</label>
             <select name={'displayLanguage'} className={'settingsContainer__select'}>
-                <option value={'de'}>deutsch</option>
-                <option value={'en'}>englisch</option>
-                <option value={'es'}>spanisch</option>
-                <option value={'gr'}>griechisch</option>
+                {
+                    displayLanguages.map(({ isSelected, languageLabel, locale }) => (
+                        <option selected={isSelected} value={locale}>{languageLabel}</option>
+                    ))
+                }
             </select>
 
             <input type={'button'} className={'settingsContainer__button'} value={'Einstellungen speichern'} />
@@ -24,7 +25,16 @@ const Settings = ({ renderModule }) => {
 }
 
 Settings.propTypes = {
-    renderModule: PropTypes.bool.isRequired
+    displayLanguages: PropTypes.arrayOf(PropTypes.shape({
+        isSelected: PropTypes.bool.isRequired,
+        languageLabel: PropTypes.string.isRequired,
+        locale: PropTypes.string.isRequired
+    })).isRequired,
+    renderModule: PropTypes.bool.isRequired,
+    translations: PropTypes.shape({
+        displayLanguageLabel: PropTypes.string.isRequired,
+        titleLabel: PropTypes.string.isRequired
+    }).isRequired
 };
 
 export default Settings;