+++ CHANGELOG.md
@@ -6,6 +6,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 ### 2020-11-09
+- Changed | bound first contents to real translations
+- Added   | reducer and default state for user settings
 - Added   | modules can now be activated by address navigation
 
 ### 2020-11-08